30 equipable (#46)
* Create equippable.py

* Spelling Error - EQUIP FABLE

* Gun / Debounce

Guns now have a fire rate variable. The keyboard controller runs a boolean function of the player's weapon slot which is currently the Gun. The Gun returns True if it can fire otherwise it returns False. The Player's weapon_slot updates everytime the player updates.

* Passing Make Types

Apparently we have Boundaries to Prevent Spaghetti Code

* made requested changes

* updated the default and create scenes

---------

Co-authored-by: Chris Palmer <capalmer1013@gmail.com>